Added VostokScheduledAsyncApplication and inherited VostokScheduledApplication from it
